restructure the whole event module
