+++ .github/workflows/crossterm_test.yml
@@ -57,7 +57,7 @@ jobs:
       run: cargo test --lib --features serde -- --nocapture --test-threads 1
       continue-on-error: ${{ matrix.can-fail }}
     - name: Test event-stream feature
-      run: cargo test --lib --features event-stream -- --nocapture --test-threads 1
+      run: cargo test --lib --features "event-stream,events" -- --nocapture --test-threads 1
       continue-on-error: ${{ matrix.can-fail }}
     - name: Test all features
       run: cargo test --all-features -- --nocapture --test-threads 1
