+++ src/event/event_api/waker/mio.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc, Mutex};
 
-use mio::{Registry, Token};
+use ::mio::{Registry, Token};
 
 use crate::Result;
 
@@ -8,7 +8,7 @@ use crate::Result;
 /// This type wraps `mio::Waker`, for more information see its documentation.
 #[derive(Clone, Debug)]
 pub(crate) struct Waker {
-    inner: Arc<Mutex<mio::Waker>>,
+    inner: Arc<Mutex<::mio::Waker>>,
 }
 
 impl Waker {
