+++ src/terminal/sys.rs
@@ -19,3 +19,7 @@ mod windows;
 
 #[cfg(unix)]
 mod unix;
+#[cfg(unix)]
+pub mod file_descriptor;
+#[cfg(unix)]
+pub mod waker;
